Add some additional headers, fix switch wiring.
Add extra I2C connector that will work with a right-angle header.
Add a neopixel header for a possible system status light.
Wire switch to input (missed that in the schematic)
Trim the number of headers for holding the XBee regulated breakout to make routing easier
Add a bunch of labels to things on the top silkscreen.